feat(svg): add a gradientRatio prop (#140)
* feat(svg): add an interval prop

Add an interval prop that controls the duration of the interval between two animations. For
instance, an interval of `.25` and an animation speed of `2` imply that the animation will
take 1.5s to complete, and wait 0.5s before starting again.

* feat(svg): add a gradientRatio prop

Add a gradientRatio prop that controls the width of the gradient relative to the viewbox's own
width. For instance, a gradientRatio of `0.5` and a width of `200` imply that the resulting gradient
will have a width of 100px.

* docs(docz): add props section